\name{bivariate0}
\alias{bivariate0}

\title{Analyse graphique bivariee par tranche}
\description{
Serie d'analyses graphiques bivariees entre un jeu de donnees et une variable specifique (explicative). Selon le type de variables 
des graphiques adaptes sont choisis.
}
\usage{
bivariate0(X,y,...)
}

\arguments{
  \item{X}{
Le nom du jeu de donnees contenant la serie de variables a expliquer}
\item{y}{La variable explicative}
\item{...}{Différents paremtres graphiques}
}


\value{
Pas de valeurs renvoyees, seulement une serie de graphiques
}


\author{
Stephane Champely
}
\note{
inspire par la fonction gpairs
}


\seealso{
gpairs
}
