\name{plotCat2}
\alias{plotCat2}

\title{Graphique pour analyse factorielle (categories)}
\description{
Realise un graphique pour des variables categorielles a partir des variables de synthese de l'analyse multivariee
}
\usage{
plotCat2(X, LI, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
Le jeu de donnees contenant les variables
}
  \item{LI}{
Les variables de synthese
}
  \item{\dots}{
Divers renseignements graphiques
}
}

\value{
Pas de valeurs renvoyees, seulement un graphique
}


\author{
Stephane Champely
}

\seealso{
plotCat, plotNum
}
