\name{plotNum}
\alias{plotNum}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Graphique pour analyse factorielle (numeriques)}
\description{
Realise un graphique pour des variables numeriques a partir des variables de synthese de l'analyse multivariee
}
\usage{
plotNum(X, LI, ...)
}


\arguments{
  \item{X}{
Le jeu de donnees
}
  \item{LI}{
Les variables de synthese
}
  \item{\dots}{
Tout autre renseignement pour la fonction graphique
}
}


\value{
Pas de valeurs renvoyees, seulement un graphique
}


\author{
Stephane Champely}


\seealso{
plotCat,plotCat2
}
