\name{RcmdrPlugin.pointG-internal}
\title{Internal RcmdrPlugin.pointG objects}




\alias{pem}

\alias{ppplotLikert0}

\alias{bbref0}
\alias{bref.df0}
\alias{bref.variable0}


\alias{sSummarydf0}
\alias{Summarydf0}
\alias{Summaryvariable0}
\alias{Summarynumeric0}
\alias{Summaryfactor0}
\alias{Summaryordered0}

\alias{ppyramide0}
\alias{pyramide0}
\alias{histbackback0}

\alias{cadrage0}

\alias{uunivariate0}

\alias{bbbivariate0}

\alias{bigraphe0}
\alias{bbigraphe0}

\alias{aaggr0}


\alias{mmultivariate0}

\alias{nettoyage0}

\alias{cchalandise.DEP0}
\alias{chalandise.DEP0}

\alias{bbbisummarize0}

\alias{llinear0}

\alias{plotTabCont0}
\alias{twoWayTable0}

\alias{ccroise0}

\alias{Yule1}
\alias{Yule2}
\alias{Yule3}
\alias{Yule4}


\description{Internal RcmdrPlugin.pointG objects.}
\details{These are not to be called by the user.}
\keyword{internal}
