\name{RcmdrPlugin.pointG-package}
\alias{RcmdrPlugin.pointG-package}
\alias{RcmdrPlugin.pointG}
\docType{package}
\title{
Install the exciting pointG Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" for studying questionnaire data.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.Gpoint\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2012-02-14\cr
License: \tab GPL (>= 2)\cr
}
}

\author{
Stephane Champely <champely@univ-lyon1.fr>

Maintainer: Stephane Champely <champely@univ-lyon1.fr>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

