\name{univariate0}
\alias{univariate0}
\title{Analyse graphique univariee}

\description{Analyse univariee graphique de toutes les variables dans un jeu de donnees. 
Selon le type de donnees, un graphique adapte est employe.
}
\usage{
univariate0(XX)
}


\arguments{
  \item{XX}{
Le nom du fichier de donnees
}
}


\value{
Pas de valeurs renvoyees, uniquement une serie de graphiques
}


\author{
Stephane Champely
}




\seealso{
whatis, summary
}

