\name{VIH}

\docType{data}

\alias{VIH}

\title{Sport et VIH}

\description{
Ce jeu de donnees est un extrait de l'enquete "Situation des personnes vivant avec le VIH dans l'acces aux activites physiques et sportives"
menee par le laboratoire Santesih (Universite Montpellier 1) et le Reseau National de personnes vivant avec le VIH.
}

\usage{data(VIH)}

\format{Un jeu de donnees avec 619 sondes et 39 variables:
Sport_Avant,Sport,Politique,Diplome,Revenu,Date_Naissance,
Poids,Taille,Sexualite,Profession,Accepte_VIH,Regards_Corps,
Corps_PasMoi,Corps_TropMaigre,Corps_TropGros,Corps_PasSeduisant,Corps_Deforme,Corps_Mieux,
Decouverte_VIH,Connait_CD4,Impact_VIH,Pensee_VIH,Evol_Loisirs,Evol_Prof,
Age,Sexe2,Sexe,DiplomeN,RevenuN,Accepte_VIHN,
Pensee_VIHN,Corps_PasMoiN,Corps_TropMaigreN,Corps_TropGrosN,Corps_PasSeduisantN,Corps_DeformeN,
Corps_MieuxN, CodePostal, Departement.
}

\source{S.Ferez et J.Thomas (Universite Montpellier 1)}


\keyword{datasets}
