\name{XY.DEP}

\docType{data}

\alias{XY.DEP}

\title{Latitude et Longitude des 95 departements en France}

\description{
Latitude et Longitude des 95 departements en France, a utiliser avec le package maps.
}

\usage{data(XY.DEP)}

\format{Un jeu de donnees avec 95 lignes et 2 colonnes : x et y.
}

\source{Personnelle}


\keyword{datasets}
