\name{plotLikert0}
\alias{plotLikert0}
\title{Analyse graphique d'echelles de Likert similaires}
\description{
Realise une analyse graphique d'une batterie d'echelles de Likert similaires, eventuellement differenciees selon
une variable categorielle.}

\usage{
plotLikert0(XX, yy=FALSE,...)
}

\arguments{
  \item{XX}{
Le nom du jeu de donnees contenant la batterie d'echelles
}
  \item{yy}{
Le nom de la variable categorielle
}
\item{...}{
Divers renseignements graphiques
}
}

\value{
Pas de valeurs renvoyees, seulement des graphiques
}

\author{
Stephane Champely
}

