\name{RcmdrPlugin.pointG-internal}
\title{Internal RcmdrPlugin.pointG objects}



\alias{vpAFC}
\alias{mapAFC}
\alias{tauK}

\alias{statMissing}

\alias{pem}
\alias{localYule}

\alias{effectPlotsTRUE}
\alias{sortAnova}


\alias{cohenW}
\alias{cramerV}


\alias{pplotLikert0}
\alias{pplotLikert1}

\alias{plotLikert1}

\alias{statLikert0}
\alias{statLikert1}

\alias{plotMissing}

\alias{bbref0}
\alias{bref.df0}


\alias{sSummarydf0}
\alias{Summarydf0}
\alias{Summarynumeric0}
\alias{Summaryfactor0}
\alias{Summaryordered0}

\alias{ppyramide0}
\alias{pyramide0}
\alias{histbackback0}

\alias{cadrage0}

\alias{uunivariate0}

\alias{bbbivariate0}

\alias{bigraphe0}
\alias{bbigraphe0}

\alias{aaggr0}


\alias{mmultivariate0}

\alias{nettoyage0}

\alias{cchalandise.DEP0}
\alias{chalandise.DEP0}

\alias{bbbisummarize0}

\alias{plotTabCont0}
\alias{twoWayTable0}

\alias{ccroise0}

\alias{Yule01}
\alias{Yule02}
\alias{Yule04}
\alias{yyule0}


\alias{enConstruction}
\alias{globals}



\description{Internal RcmdrPlugin.pointG objects.}
\details{These are not to be called by the user.}
\keyword{internal}
